Refactor and restructure DVLD solution
Refactored `ApplicationTypesController` to improve parameter naming and added a `GetFees` endpoint. Updated `IApplicationTypesRepository` and `ApplicationTypesService` to handle nullable application fees. Modified `DvldApplicationsService` to cast fees appropriately.

Removed `DriversController`, `PeopleController`, and `UsersController`, along with their associated DTOs, services, and repositories. Deleted `DVLD.API`, `DVLD.Application`, `DVLD.Domain`, and `DVLD.Infrastructure` project configurations, including `DvldDbContext` and all entity classes.

Eliminated the solution file (`DVLD.sln`) and performed general cleanup by removing unused classes and mappers. These changes reflect a significant architectural overhaul and a shift toward a streamlined design.

+++ DVLD.Application/Services/DvldApplicationsService.cs
@@ -179,7 +179,7 @@ public class DvldApplicationsService
                 request.ApplicantPersonId,
                 request.ApplicationTypeId,
                 request.ApplicationStatus,
-                _applicationTypesRepository.GetApplicationFees(request.ApplicationTypeId).Result,
+                (decimal)_applicationTypesRepository.GetApplicationFees(request.ApplicationTypeId).Result,
                 request.CreatedByUserId);
